* 멤버 조회 시 학과명 추가
* 멘토링 리스트 조회 시 날짜/시간 순 정렬
* 공지사항 수정 시 일반 사용자에게 안보이던 버그 수정
* 날짜 충돌 로직 변경 (날짜 같으면 바로 충돌)

+++ src/main/java/pheonix/classconnect/backend/security/utils/PrincipalUtils.java
@@ -7,10 +7,10 @@ import pheonix.classconnect.backend.exceptions.MainApplicationException;
 
 @Component
 public class PrincipalUtils {
-    public static Integer getUserId(User user) {
+    public static Long getUserId(User user) {
         if (user == null) {
             throw new MainApplicationException(ErrorCode.BACK_INVALID_PERMISSION, "인증객체에서 유저 정보를 찾을 수 없습니다.");
         }
-        return Integer.parseInt(user.getUsername());
+        return Long.parseLong(user.getUsername());
     }
 }
